fix(quota): raise the quota ceiling and widen quota columns to bigint
Saving an organization quota failed with "quota must be between 0 and
1000000000". That ceiling was only $2,000 at the default QuotaPerUnit, far too
low -- but raising it past 1e9 exceeds a 32-bit column, so the ceiling and the
storage have to move together.

Introduces common.MaxQuota = 5e14 (exactly $1B at the default QuotaPerUnit) as
the single definition, referenced by controller.MaxOrganizationQuota. 5e14 keeps
~18x headroom under JavaScript's Number.MAX_SAFE_INTEGER, so quota stays exactly
representable in the browser -- including when the quota input multiplies a typed
amount by QuotaPerUnit, and when the dashboard sums several balances.

Widens to int64 / gorm:"type:bigint":
  - Organization.Quota/UsedQuota and the dashboard rows
  - User.Quota/UsedQuota/AffQuota/AffHistoryQuota
  - WalletAutoRecharge(.Preset).ThresholdQuota

The user columns are not incidental. The organization user table edits a member's
personal quota, which lives in users.quota -- not organizations.quota -- and that
field is validated against MaxOrganizationQuota. Leaving it 32-bit meant
PostgreSQL raised "integer out of range", non-strict MySQL silently clamped, and
only SQLite worked: a value that persists on one supported database and fails on
another, violating CLAUDE.md rule 2. AffQuota moves too because
TransferAffQuotaToQuota shifts value directly between it and Quota in one
transaction.

Org and user widening are one commit because they are inseparable: the dashboard
rows, logger.FormatQuota and the service layer all flip types together, so any
split leaves a commit that does not compile.

The cascade is mechanical -- the compiler drove it -- and it removes narrowing
casts rather than adding them. The quota-unit settings (QuotaForNewUser/Inviter/
Invitee, QuotaRemindThreshold) widen with it, so model/option.go parses them with
ParseInt instead of Atoi. Token.RemainQuota and Channel.UsedQuota stay int -- a
separate unit domain -- and every crossing point wraps explicitly.

Two things deliberately did NOT move to the new ceiling:
  - Top-up amounts get their own MaxTopUpAmount, keeping the original 1e9. They
    are currency units handed to a payment gateway and later multiplied by
    QuotaPerUnit; sharing the quota ceiling would have quietly widened a payment
    guard by six orders of magnitude.
  - The Toss per-payment math.MaxInt32 bound stays. It is a payment policy, not a
    storage limit; the stale comment claiming it reflects a SQL INT column is
    corrected.

The ceiling is enforced by request validation and, for payments, by
CreditTopUpTarget's existing atomic guard (now bounding against MaxQuota rather
than math.MaxInt32). It is deliberately NOT enforced on the generic increment
helpers: most callers of IncreaseUserQuota are refunds and compensations, and a
refund that cannot be applied destroys quota the account already paid for --
worse than the precision drift a ceiling there would prevent. The Redis quota
cache is also credited before the DB write, so a rejected write would leave
spendable phantom quota.

Fixes one latent bug found on the way: ContextKeyUserQuota is now written as
int64, which common.GetContextKeyInt would have silently read back as 0. Adds
GetContextKeyInt64 and switches the single reader.

MaxOrganizationQuota is typed int64 so the untyped constant cannot default to
`int` and overflow a 32-bit build. Test expectations derive from the constant
instead of hardcoding the number.

AutoMigrate performs the int -> bigint widening on all three databases with no
manual DDL; verified on SQLite that existing rows survive the table rebuild
byte-exact and that the rebuild is transactional.

+++ controller/checkin.go
@@ -61,7 +61,7 @@ func DoCheckin(c *gin.Context) {
 		})
 		return
 	}
-	model.RecordLog(userId, model.LogTypeSystem, fmt.Sprintf("Check-in successful, quota awarded: %s", logger.LogQuota(checkin.QuotaAwarded)))
+	model.RecordLog(userId, model.LogTypeSystem, fmt.Sprintf("Check-in successful, quota awarded: %s", logger.LogQuota(int64(checkin.QuotaAwarded))))
 	c.JSON(http.StatusOK, gin.H{
 		"success": true,
 		"message": common.TranslateMessage(c, i18n.MsgCheckinSuccess),
